type: float to double

+++ src/lang/type/t.c
@@ -6,8 +6,8 @@ main(int argc, const char *argv[]) {
 	_unused_(argc);
 	_unused_(argv);
 
-  float f0 = 3.14;
-  float f1 = 2.7182;
+  float f0 = 3.14f;
+  double f1 = 2.7182;
   printf("pi %f stored as %8x\n", f0, *(unsigned int *)&f0);
-  printf("e %f stored as %8x\n", f1, *(unsigned int *)&f1);
+  printf("e %f stored as %8x\n", f1, *(unsigned long*)&f1);
 }
